refactor: design 폴더 생성 및 MyPageSection 폴더명 변경
