+++ src/pages/MyPage/HistoryPage/TabPage/PostsPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import PageNavigation from '../../../../components/Paginavigation/Paginavigation';
-import List from '../../MypageSections/List/List';
-import ItemPosts from '../../MypageSections/ListItem/ItemPosts';
+import List from '../../sections/List/List';
+import ItemPosts from '../../sections/ListItem/ItemPosts';
 
 const PostsPage = () => {
   const items = [
