+++ src/pages/MyPage/sections/ListItem/ItemVotedPosts.tsx
