+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Button from '../../components/common/Button/Button';
-import Heading from '../../components/common/Heading/Heading';
+import Button from '../../components/design/Button/Button';
+import Heading from '../../components/design/Heading/Heading';
 import InputCode from '../../components/InputsUserInfo/InputCode';
 import InputEmail from '../../components/InputsUserInfo/InputEmail';
 import InputNickname from '../../components/InputsUserInfo/InputNickname';
